move Extrude and test into function package
